% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dof.R
\name{dof_pladj}
\alias{dof_pladj}
\title{Degrees of freedom using the partial leverages adjustment proposed by Kranz (2024)}
\usage{
dof_pladj(x)
}
\arguments{
\item{x}{A regression result from lm or feols. Alternatively, a matrix of partial leverage generated with partial_leverages.}
}
\description{
Use these degree of freedoms for the t-test of HC1 or HC2 standard errors
Only makes sense for these hetereoksedasticity robust standard errors.
Does not work for cluster robust standard errors
}
